+++ lldb/include/lldb/Host/common/DiagnosticsRendering.h
