+++ lldb/unittests/Host/common/CMakeLists.txt
@@ -1,4 +1,5 @@
 set (FILES
+  DiagnosticsRenderingTest.cpp
   ZipFileResolverTest.cpp
 )
