+++ llvm/utils/gn/secondary/lldb/source/Host/BUILD.gn
@@ -17,6 +17,7 @@ static_library("Host") {
   public_deps = [ "//llvm/utils/gn/build/libs/xml" ]
   sources = [
     "aix/Support.cpp",
+    "common/DiagnosticsRendering.cpp",
     "common/File.cpp",
     "common/FileAction.cpp",
     "common/FileCache.cpp",
