+++ llvm/utils/gn/secondary/lldb/source/Utility/BUILD.gn
@@ -21,7 +21,6 @@ static_library("Utility") {
     "DataEncoder.cpp",
     "DataExtractor.cpp",
     "Diagnostics.cpp",
-    "DiagnosticsRendering.cpp",
     "Environment.cpp",
     "ErrorMessages.cpp",
     "Event.cpp",
